plugins/rancher/pyproject.toml
[project]
name = "waldur-site-agent-rancher"
version = "0.7.0"
description = "Rancher plugin for Waldur Site Agent"
readme = "README.md"
authors = [{ name = "OpenNode Team", email = "info@opennodecloud.com" }]
requires-python = ">=3.9, <4"
dependencies = [
    "waldur-site-agent>=0.7.0",
    "requests>=2.31.0",
    "python-keycloak>=3.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
waldur-site-agent = { workspace = true }

# Entry points for exporting backends
[project.entry-points."waldur_site_agent.backends"]
rancher = "waldur_site_agent_rancher.backend:RancherBackend"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
]
